+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/MetroOptions.kt
@@ -470,7 +470,7 @@ internal enum class MetroOption(val raw: RawMetroOption<*>) {
   ENABLE_SCOPED_INJECT_CLASS_HINTS(
     RawMetroOption.boolean(
       name = "enable-scoped-inject-class-hints",
-      defaultValue = true,
+      defaultValue = false,
       valueDescription = "<true | false>",
       description =
         "Enable/disable generating hints for scoped @Inject classes. By default, a scoped injectable class that isn't used in its associated graph node will result in an error if a graph extension later tries to inject it. Enabling this setting prevents such errors by generating a binding for all scoped types within the graph node. See https://github.com/ZacSweers/metro/issues/377 for more context.",