+++ compiler/src/test/kotlin/dev/zacsweers/metro/compiler/ir/TracingTest.kt
@@ -41,8 +41,8 @@ class TracingTest : MetroCompilerTest() {
       assertThat(withoutTime)
         .isEqualTo(
           """
-          main,Collecting contributions
           main,Transforming Metro contributions
+          main,Collecting contributions
           ExampleGraph,Build DependencyGraphNode
           ExampleGraph,Implement creator functions
           ExampleGraph,Build binding graph
@@ -77,10 +77,10 @@ class TracingTest : MetroCompilerTest() {
         .isEqualTo(
           """
             [main] ▶ Metro compiler
-              ▶ Collecting contributions
-              ◀ Collecting contributions (xx ms)
               ▶ Transforming Metro contributions
               ◀ Transforming Metro contributions (xx ms)
+              ▶ Collecting contributions
+              ◀ Collecting contributions (xx ms)
               ▶ Core transformers
                 ▶ Transform dependency graph
                   ▶ Build DependencyGraphNode
@@ -173,8 +173,8 @@ class TracingTest : MetroCompilerTest() {
       assertThat(withoutTime)
         .isEqualTo(
           """
-          main,Collecting contributions
           main,Transforming Metro contributions
+          main,Collecting contributions
           ExampleGraph,Build DependencyGraphNode
           ExampleGraph,Implement creator functions
           ExampleGraph,Build binding graph
@@ -232,10 +232,10 @@ class TracingTest : MetroCompilerTest() {
         .isEqualTo(
           """
             [main] ▶ Metro compiler
-              ▶ Collecting contributions
-              ◀ Collecting contributions (xx ms)
               ▶ Transforming Metro contributions
               ◀ Transforming Metro contributions (xx ms)
+              ▶ Collecting contributions
+              ◀ Collecting contributions (xx ms)
               ▶ Core transformers
                 ▶ Transform dependency graph
                   ▶ Build DependencyGraphNode
@@ -374,8 +374,8 @@ class TracingTest : MetroCompilerTest() {
       assertThat(withoutTime)
         .isEqualTo(
           $$$"""
-          main,Collecting contributions
           main,Transforming Metro contributions
+          main,Collecting contributions
           ExampleGraph,Build DependencyGraphNode
           ExampleGraph,Implement creator functions
           ExampleGraph,Build binding graph
@@ -434,10 +434,10 @@ class TracingTest : MetroCompilerTest() {
         .isEqualTo(
           $$$"""
             [main] ▶ Metro compiler
-              ▶ Collecting contributions
-              ◀ Collecting contributions (xx ms)
               ▶ Transforming Metro contributions
               ◀ Transforming Metro contributions (xx ms)
+              ▶ Collecting contributions
+              ◀ Collecting contributions (xx ms)
               ▶ Core transformers
                 ▶ Transform dependency graph
                   ▶ Build DependencyGraphNode