+++ compiler/src/test/kotlin/dev/zacsweers/metro/compiler/transformers/ContributesGraphExtensionTest.kt
@@ -1214,7 +1214,8 @@ class ContributesGraphExtensionTest : MetroCompilerTest() {
           }
         """
           .trimIndent()
-      )
+      ),
+      options = metroOptions.copy(enableScopedInjectClassHints = true),
     ) {
       val parentGraph = ExampleGraph.generatedMetroGraphClass().createGraphWithNoArgs()
       val childGraph = parentGraph.callFunction<Any>("createLoggedInGraph")
@@ -1249,7 +1250,8 @@ class ContributesGraphExtensionTest : MetroCompilerTest() {
           }
         """
           .trimIndent()
-      )
+      ),
+      options = metroOptions.copy(enableScopedInjectClassHints = true),
     ) {
       val parentGraph = ExampleGraph.generatedMetroGraphClass().createGraphWithNoArgs()
       val childGraph = parentGraph.callFunction<Any>("createLoggedInGraph")
@@ -1284,7 +1286,8 @@ class ContributesGraphExtensionTest : MetroCompilerTest() {
           }
         """
           .trimIndent()
-      )
+      ),
+      options = metroOptions.copy(enableScopedInjectClassHints = true),
     ) {
       val parentGraph = ExampleGraph.generatedMetroGraphClass().createGraphWithNoArgs()
       val childGraph = parentGraph.callFunction<Any>("createLoggedInGraph")
@@ -1306,7 +1309,8 @@ class ContributesGraphExtensionTest : MetroCompilerTest() {
           class Dependency : Bob
         """
             .trimIndent()
-        )
+        ),
+        options = metroOptions.copy(enableScopedInjectClassHints = true),
       )
 
     val graphExtensionCompilation =
@@ -1327,6 +1331,7 @@ class ContributesGraphExtensionTest : MetroCompilerTest() {
         """
             .trimIndent()
         ),
+        options = metroOptions.copy(enableScopedInjectClassHints = true),
         previousCompilationResult = injectDepCompilation,
       )
 
@@ -1342,6 +1347,7 @@ class ContributesGraphExtensionTest : MetroCompilerTest() {
         addPreviousResultToClasspath(injectDepCompilation)
         addPreviousResultToClasspath(graphExtensionCompilation)
       },
+      options = metroOptions.copy(enableScopedInjectClassHints = true),
     ) {
       val parentGraph = ExampleGraph.generatedMetroGraphClass().createGraphWithNoArgs()
       val childGraph = parentGraph.callFunction<Any>("createLoggedInGraph")
@@ -1362,7 +1368,8 @@ class ContributesGraphExtensionTest : MetroCompilerTest() {
           class Dependency
         """
             .trimIndent()
-        )
+        ),
+        options = metroOptions.copy(enableScopedInjectClassHints = true),
       )
 
     val graphExtensionCompilation =
@@ -1384,6 +1391,7 @@ class ContributesGraphExtensionTest : MetroCompilerTest() {
             .trimIndent()
         ),
         previousCompilationResult = injectDepCompilation,
+        options = metroOptions.copy(enableScopedInjectClassHints = true),
       )
 
     compile(
@@ -1399,6 +1407,7 @@ class ContributesGraphExtensionTest : MetroCompilerTest() {
         addPreviousResultToClasspath(injectDepCompilation)
         addPreviousResultToClasspath(graphExtensionCompilation)
       },
+      options = metroOptions.copy(enableScopedInjectClassHints = true),
     ) {
       val parentGraph = ExampleGraph.generatedMetroGraphClass().createGraphWithNoArgs()
       val childGraph = parentGraph.callFunction<Any>("createLoggedInGraph")
@@ -1442,7 +1451,8 @@ class ContributesGraphExtensionTest : MetroCompilerTest() {
           }
         """
           .trimIndent()
-      )
+      ),
+      options = metroOptions.copy(enableScopedInjectClassHints = true),
     ) {
       val parentGraph = ExampleGraph.generatedMetroGraphClass().createGraphWithNoArgs()
       val childGraph1 = parentGraph.callFunction<Any>("createLoggedInGraph")