+++ gradle-plugin/src/main/kotlin/dev/zacsweers/metro/gradle/MetroPluginExtension.kt
@@ -68,13 +68,13 @@ constructor(layout: ProjectLayout, objects: ObjectFactory, providers: ProviderFa
     objects.property(Boolean::class.javaObjectType).convention(false)
 
   /**
-   * Enable/disable hint property generation for scoped inject classes. Enabled by default.
+   * Enable/disable hint property generation for scoped inject classes. Disabled by default.
    *
    * @see <a href="https://zacsweers.github.io/metro/dependency-graphs/#graph-extensions">Graph
    *   Extensions docs for more details</a>
    */
   public val enableScopedInjectClassHints: Property<Boolean> =
-    objects.property(Boolean::class.javaObjectType).convention(true)
+    objects.property(Boolean::class.javaObjectType).convention(false)
 
   /** Enable/disable shrinking of unused bindings. Enabled by default. */
   public val shrinkUnusedBindings: Property<Boolean> =